Android: Get rid of init classes mechanism
Make it more the responsibility of the module to ensure
any initialization is done under JNI_OnLoad().

Qt for Android have these mechanism to set the context to Java static
classes used by various modules, where the QtLoader tries to look for
setContext(Context), setActivity(Activity) and setService(Service)
methods defined under certain modules where it uses reflection to,
for example, assign the context to them before actually loading their
respective shared libraries where some initialization is needed and
requires the context to be already set.

This mechanism is not really necessary, since the respective libraries
are going to be loaded first anyways, and they could ensure to init
the context, if needed, under JNI_OnLoad() instead of those static
classes and through reflection which requires the class to be public.
Also, this way it would be more explicit behavior where the context is
set exactly before it's needed by the module's library.

Fixes: QTBUG-126478
Change-Id: I4cb39b8ef057ff7a99b66c268d85ba6da0c591ce
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>