d3d: Drive window updates from a vblank watcher thread
Have dedicated threads that just do IDXGIOutput::WaitForVBlank() and
fire off notifications. I.e., create something conceptually similar to
macOS' CVDisplayLink. Then hook this up to QWindow::requestUpdate().

For windows that are not D3D and not used with an active QRhi, or when
the used D3D adapter is WARP that reports no DXGI outputs, this has no
effect, and everything works like before, with the standard
requestUpdate behavior of using a 5 ms (or less) timer.

For development purposes one can always disable this by setting the
environment variable QT_D3D_NO_VBLANK_THREAD to a non-zero value. To
have some debug output, enable the category qt.qpa.screen.updates.

Multiple refresh rates are supported by using dedicated a notifier
thread per output. This way we do not bake in assumptions about the
composition and presentation behavior, which is good since it is
reported to have changed around Win 11 23H2 (DWM not anymore limiting to
the primary display's refresh rate). E.g., just having a single thread
ticking based on the primary screen would be quite wrong.

The result is a very visibly reduced display lag e.g. with the classic
test case of dragging an item with the mouse around in a Qt Quick scene,
less time spent on throttling to the presentation rate (and more time in
the event loop, presumably), and, for some reason, a noticeable
reduction in the reported CPU load as well.

This complements the 6.8 changes that added a blocking wait on the frame
latency waitable object (with max latency set to 2) in beginFrame().
Those changes provided some improvements for the lag-when-dragging case
in particular, effectively bringing the out of the box Qt 6 experience
with D3D on par with the Qt 5 results with OpenGL (results vary with
diferent vendors; all testing here has been done with NVIDIA graphics).
Whereas this more complex approach on top provides further improvements,
which were not possible to achieve before with any configuration of 3D
APIs or settings.

[ChangeLog][Windows] By default, windows that get rendered to with QRhi
using either the D3D11 or D3D12 backend base their update request
deliveries on notifications from dedicated vblank waiting threads. This
affects both direct users of QWindow::requestUpdate() and also Qt Quick,
since QQuickWindow uses the same mechanism internally. This is expected
to provide smoother presentation, in particular a reduced display -
mouse cursor lag when dragging items in a Qt Quick scene for example,
and potentially reduced CPU load. Note that this is not applicable when
using an adapter such as WARP. For development and testing purposes, it
is possible to request the old behavior by setting the environment
variable QT_D3D_NO_VBLANK_THREAD to a non-zero value.

Task-number: QTBUG-127267
Change-Id: I6e35e776f495bc34cf164164da8ce7f9e74f45b6
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>