tst_QStringBuilder: add a check for QBA/QL1SV concatenation
Rvalue QL1S and (const) lvalue QBA combination is ambiguous in
qtlottie tests on the latest qt5.git integration round.

It was not caugth by the QStringBuilder tests, because the
tests never checked op+() with allowed casts to/from ASCII.

Add a new test scenario and hide the failing combination
under ifdef (cannot use QEXPECT_FAIL as it's a compile-time
error).
The ifdef will be removed in a follow-up patch that fixes
the issue.

Task-number: QTBUG-127904
Task-number: QTBUG-127928
Task-number: QTBUG-127931
Change-Id: I92b582f16e850fb34c34b75714d820e65d411265
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>