QMacMimeRtfText: Remove stray newline when converting RTF to HTML
NSAttributedString wrongly inserts a newline at the end of the string
when generating HTML from RTF. The rules for HTML parsing specify that
this newline, after the closing </html> tag, should be treated the same
way as if it had been part of the body.

  - https://html.spec.whatwg.org/multipage/syntax.html#start-tags
  - https://html.spec.whatwg.org/multipage/parsing.html#the-after-after-body-insertion-mode

This results in an additional space character if the HTML is converted
to a QTextDocument by QTextHtmlParser, for example when inserting text
from another application via the clipboard or drag-and-drop.

Pick-to: 6.8 6.5 6.2 5.15
Change-Id: If17cfcdc9ad5452e7b516fc6b50d834bd5184e23
Reviewed-by: Michael Brüning <michael.bruning@qt.io>