Record the CLDR bug tracking the mn_Mong_MN number format
As previously commented, mn_Mong_MN would end up with the same decimal
and group separators if we trusted its draft="contributed" amendment
to the decimal separator. I reported this to the CLDR folk some time
ago and they now have a Jira ticket for it, which turns out to be a
duplicate, so we can track them and know when to remove my
hack-around.

Change-Id: Ib8f49dbdce090393ad20cd50969d6323818ee4ff
Reviewed-by: Mate Barany <mate.barany@qt.io>