Eliminate a potential bug
If we send an event here, we may enter a new event loop and then
the new event loop will process the events including socket events.
As a result, after exiting the new event loop, the activeNotifierPos
is same as the src->pollfds.count() and activeNotifierPos will be
added by one in the old event loop.

So there is a potential situation that activeNotifierPos will overflow
if we have too many sockets and nested event loops.

Although this situation is almost impossible, we can improve the code
by little effort. That's it.