QCompleter: Forward mouse press/release to virtual keyboard
When the list view of a QCompleter is visible, it receives mouse events.
A mouse press event outside the completer hid it.

If a virtual keyboard was active, it did not receive mouse press/release
events in case the completer's list view was active.

Add a helper to QGuiApplicationPrivate, that detects a virtual keyboard
under a mouse event position, and forwards the event if needed.

Adapt QCompleter::eventFilter() to use the helper.
If a virtual keyboard one exists and a mouse press event occurs inside
it, don't hide the completer and forward the event.
Add handling of mouse release events, and also forward them to the
virtual keyboard, using the helper.

Add logging.

Fixes: QTBUG-126752
Pick-to: 6.8 6.7 6.5 6.2
Change-Id: I613523a9c83b319810397770c43b2eb4ae01c31a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>