QEventDispatcherGlib: avoid potential integer overflow
There is a fact that the value of sizeof(int) may be smaller than the
value of sizeof(qsizetype).

Now that the data type of the return value for function QList::size() is
qsizetype, we should change the datatype of activeNotifierPos from int
to qsizetype in case that it may overflow in the function
socketNotifierSourceDispatch.

And we should also change the data type of the iteration variable i from int
to qsizetype for the same reason in the function unregisterSocketNotifier.

Change-Id: I017d184aa046684c0af1c17c94fb83611706ffa3
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>