moc: add support for storing 64-bit enums and flags
The metatype format is extended by storing the high parts of all
enumerations after the (name, low part) section, for a total of 12 bytes
per enumeration instead of 8. This allows 32-bit flags and enums to
remain unchanged. In fact, older meta object parsers remain unchanged,
aside from not knowing what this new bit in the flags field means.

Because this is entirely done in the C++ constexpr side, moc knows
nothing about it so it can't update the JSON output with a new field.

Task-number: QTBUG-111926
Change-Id: I8a96935cf6c742259c9dfffd17e9512e4f4add19
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>